+++ STARTER_PCSS/gulpfile.js
@@ -2,6 +2,7 @@ const { src, dest, watch, parallel } = require('gulp');
 const sourcemaps = require('gulp-sourcemaps');
 const postcss = require('gulp-postcss');
 const rename = require('gulp-rename');
+const terser = require('gulp-terser');
 const autoprefixer = require('autoprefixer');
 const cssnano = require('cssnano');
 const postcssCustomMedia = require('postcss-custom-media');
@@ -44,5 +45,31 @@ function watchCss() {
   watch('assets/pcss/**/*.pcss', parallel(compileCss));
 }
 
+/**
+ * Builds JS files.
+ */
+function compileJs() {
+  return src([
+    'assets/js/**/*.js',
+  ])
+    .pipe(terser())
+    .pipe(rename(function (path) {
+      // If file contains .es6.js ending, strip 'es6' part.
+      path.basename = path.basename.replace('.es6', '');
+    }))
+    .pipe(dest(destinationToSource));
+}
+
+/**
+ * Watch for JS file changes and build if any of them changed.
+ */
+function watchJs() {
+  return watch('assets/js/**/*.js', parallel(compileJs));
+}
+
+exports.compileJs = compileJs;
+exports.watchJs = watchJs;
 exports.compileCss = compileCss;
-exports.default = exports.watchCss = watchCss;
+exports.watchCss = watchCss;
+exports.watchAll = parallel(watchJs, watchCss);
+exports.default = exports.compileAll = parallel(compileJs, compileCss);